Improve private state property checks
When subscribing/unsubscribing, all events for either a student or a course
will be loaded on reconsititution. This also means subscribing events for
a student on other courses, and subscribing events for a course for other
students.

When this has happened before the course and student for the current
subscribe/unsubscribe, certain private state properties are not set yet
(as the required created events did not happen yet for this domain id combination).

This will result in uninitialized errors. Therefore, make sure this cannot
happen by setting defaults.